Rename legacy TMA info metrics
A TMA metric like Backend_Bound is renamed to tma_backend_bound, but
Info metrics Page_Walks_Utilization were not. To avoid this
inconsistency and to align with these metrics being added to Valkyrie,
lower case the metric names and prefix with 'tma_info_'. As there are
some fixes for metrics based on names, this is updated. The ignored
metrics are also changed from a set to a dictionary, so that the
before and after name can be used.

In doing this it was observed that tma_backend_bound wasn't having the
'NO_NMI_WATCHDOG' constraint added after TMA metrics were
renamed. This is also fixed.