Add event grouping flags to metrics
Linux perf opens performance counters on a fake CPU. If too many
counters are requested the open should fail but short comings in the
Linux kernel and because of errata items means this doesn't always
happen. If the fail doesn't happen the events aren't ungrouped and the
metric doesn't return a count. A command line option of
--metric-no-group can work around this, but is inconvenient. Add flags
to indicate the different situations events in metrics shouldn't be
grouped.