Parse the TMA metric Threshold column
Metric thresholds are like "> 0.1 & P" where the "> 0.1" means the
metric is over the threshold when greater-than 10% while the "& P"
means the parent metric's threshold also needs to be over the
threshold. Translate threshold formulas to be the same as the regular
formulas, removing knowledge in the process about parent nodes.

For example, tma_fetch_latency has a thresold of "> 0.1 & P" and the
metric now gets the following added:
  "MetricThreshold": "tma_fetch_latency > 0.1 & tma_frontend_bound > 0.15",

A more complex example is tma_info_dsb_coverage that from
"< 0.7 & #HighIPC; $issueFB" gets a threshold expression of:
  "MetricThreshold": "tma_info_dsb_coverage < 0.7 & tma_info_ipc / 6 > 0.35"