readme: Check for broken links
Check markdown content for broken links. A few were broken when
previously updating the Errata information section. This initial
Action is based on the recommended getting started [1]. See [2] for
details on the underlying tool wrapped by the GitHub Action. The GitHub
action wrapper [1] is MIT licensed. Markdown-link-check is ISC licensed
[2].

[1] https://github.com/gaurav-nelson/github-action-markdown-link-check#how-to-use
[2] https://github.com/tcort/markdown-link-check