Don't rewrite metrics across PMUs
Metrics are rewritten in terms of others, in part to allow the
injection of #EBS_Mode support, but also to shorten metrics and
improve readability. For hybrid it is possible that metrics for
cpu_atom are rewritten to use metrics for cpu_core, which doesn't make
sense. To avoid this, perform the rewriting per CSV file rather than
before writing.