Make events in metrics for hybrid PMU specific
An event like topdown-fe-bound exists for both PMUs and so specifying
it in the metric will cause the event to be opened on both PMUs. This
happens even though the metric is specific to a single PMU. Change
cpu_core/cpu_atom events to explicitly state the PMU in the
metric. Record the PMU name during event parsing, even though its not
part of the input json events.