Raise an exception if Valkyrie metrics aren't found
The Valkyrie metrics may or may not be present for a model, as such we
ignore exceptions in trying to open them. This has been error prone
due to file paths changing. Rather than ignore the exception, re-throw
it for the cases where we expect Valkyrie metrics to be present.