Add Default metric group for models with perf metric events
Models without perf metric events may lead to inaccuracies when run with
a single benchmark. Add a new metric group 'Default' to be used in place
of 'TopdownL1' and 'TopdownL2' by perf when adding default metrics.

Add DefaultMetricgroupName so perf can understand the real metric group
when using the default mode.

Avoiding grouping events for the default mode as well.

Signed-off-by: Kan Liang <kan.liang@linux.intel.com>