For alderlake/hybrid, add PMU for events at the start of the line
Events in a formula on alderlake have the PMU added, but care is done
to not add this twice by checking the preceeding character isn't an
'@' the PMU separator as '/' looks like a divide in formulas. This is
failing when the event has nothing before it, ie at the start of a
line, and the metric is more than just the event name - something that
already had a special case. Update the logic to handle events at the
start of a formula.

This fixes the metric logic in perf where events appearing more than
once aren't added multiple times. Previously as the names were
distinct due to a lack of PMU prefix, the same event would be added
more than once for some metrics on alderlake/hybrid.