For Linux perf converter, add EBS_Mode back to CORE_CLKs
Add EBS_Mode back to CORE_CLKs for pre-Icelake architectures. Linux
perf can be used per-thread/process or system wide. The EBS_Mode
correction to CORE_CLKS is necessary for perf's per-thread/process
numbers to be more correct, albeit the best solution is to gather the
metrics system wide.

In the future, if we want to disable CORE_CLKS based metrics in perf
then we'll still need EBS_Mode and use it to substitute a value like
NaN in. Support for this hasn't yet been added to the perf tool:
https://lore.kernel.org/lkml/20240209204947.3873294-2-irogers@google.com/
So for now keep EBS_Mode as it was.