create_perf_json: Expose PDIST Counter information
The generic PDIST availability information can only be found in the SDM.
Also, some events may not support PDIST. It's hard to find that
information either.

Expose the PDIST information in the event list.

The new field PDISTCounter indicates the event's available PDIST
counters. '0' is valid counter index.
For the events which don't support PDIST, "NA" is filled.

Signed-off-by: Kan Liang <kan.liang@linux.intel.com>