create_perf_json: Expose PDIST Counter information in PublicDescription
The generic PDIST availability information can only be found in the SDM.
Also, some events may not support PDIST. It's hard to find that
information either.

Append the PDIST information into the PublicDescription of the event
list, if it's available.

If the event or platform doesn't support PDIST, nothing is changed.

The PDIST Counter information may be useful for the hardware-aware
grouping of Linux perf. But the hardware-aware grouping only works for
perf stat for now. There is no plan to enable the hardware-aware
grouping for perf record shortly. Adding a dedicated PDIST field is
unnecessary.

Signed-off-by: Kan Liang <kan.liang@linux.intel.com>