Update metric constraints
The metric constraints will break groups in certain situations to
avoid perf not being able to program the events, typically because of
the perf event NMI watchdog. Update the constraints by logging onto
machines of each type, disabling the :W (weak group) option in perf's
metricgroup.c, then seeing which metrics can be programmed with and
without the NMI watchdog enabled.

For Alderlake, add a new option for with and without the threshold
enabled, as some metrics fail with the NMI watchdog because of the
extra events needed for the threshold.